Add swagger complements
