+++ post/api/category.py
@@ -1,3 +1,5 @@
+from drf_yasg import openapi
+from drf_yasg.utils import swagger_auto_schema
 from rest_framework import viewsets
 from rest_framework.request import Request
 from rest_framework.response import Response
@@ -7,6 +9,9 @@ from post.serializers import CategorySerializer
 
 
 class CategoryAPI(viewsets.ViewSet):
+    list_response = openapi.Response('Success', CategorySerializer(many=True))
+
+    @swagger_auto_schema(responses={200: list_response})
     def list(self, request: Request) -> Response:
         """
         Get the list of categories.
