+++ post/serializers/comment.py
@@ -1,6 +1,7 @@
 from rest_framework import serializers
 
-from post.models import Comment
+from post.models import Comment, Post
+from user.models import User
 
 
 class CommentSerializer(serializers.ModelSerializer):
@@ -35,3 +36,35 @@ class CommentPutSerializer(serializers.ModelSerializer):
     class Meta:
         model = Comment
         fields = ('id', 'content')
+
+
+class CommentCreateBodySerializer(serializers.Serializer):
+    content = serializers.CharField(
+        required=True,
+        help_text='The content of the comment',
+    )
+    post = serializers.PrimaryKeyRelatedField(
+        required=True,
+        queryset=Post.objects.all(),
+        help_text='The post ID of the comment',
+    )
+    author = serializers.PrimaryKeyRelatedField(
+        required=True,
+        queryset=User.objects.all(),
+        help_text='The author ID of the comment',
+    )
+
+    class Meta:
+        ref_name = None
+        fields = ('content', 'post', 'author')
+
+
+class CommentBodySerializer(serializers.Serializer):
+    content = serializers.CharField(
+        required=True,
+        help_text='The content of the comment',
+    )
+
+    class Meta:
+        ref_name = None
+        fields = ('content')
